[#3148] Move TelemetryExecutionContext to adapter-base module
