+++ adapters/http-vertx-base/src/test/java/org/eclipse/hono/adapter/http/HonoBasicAuthHandlerTest.java
@@ -29,6 +29,7 @@ import java.util.Base64;
 import java.util.HashMap;
 import java.util.Map;
 
+import org.eclipse.hono.adapter.HttpContext;
 import org.eclipse.hono.adapter.auth.device.AbstractDeviceCredentials;
 import org.eclipse.hono.adapter.auth.device.DeviceCredentialsAuthProvider;
 import org.eclipse.hono.adapter.auth.device.PreCredentialsValidationHandler;
@@ -36,7 +37,6 @@ import org.eclipse.hono.adapter.auth.device.UsernamePasswordCredentials;
 import org.eclipse.hono.client.ServerErrorException;
 import org.eclipse.hono.client.ServiceInvocationException;
 import org.eclipse.hono.service.auth.DeviceUser;
-import org.eclipse.hono.service.http.HttpContext;
 import org.eclipse.hono.test.VertxMockSupport;
 import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentCaptor;
